// src/lib/okane-finance-api/generated/models/BacktestStats.ts
export interface BacktestStats {
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	ticker: any | null;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	maxDrawdownPercentage: number;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	startTime: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	endTime: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	duration: any | null;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	exposureTimePercentage: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	finalEquity: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	peakEquity: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	returnPercentage: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	buyAndHoldReturn: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	returnAnnualized: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	volatilityAnnualized: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	sharpeRatio: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	sortinoRatio: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	calmarRatio: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	averageDrawdownPercentage: number;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	maxDrawdownDuration: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	averageDrawdownDuration: any | null;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	tradeCount: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	winRate: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	bestTrade: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	worstTrade: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	avgTrade: number;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	maxTradeDuration: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	averageTradeDuration: any | null;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	profitFactor: number;
	/**
	 *
	 * @type {any}
	 * @memberof BacktestStats
	 */
	html: any | null;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	tpslRatio: number;
	/**
	 *
	 * @type {number}
	 * @memberof BacktestStats
	 */
	slCoef: number;
}